adds pagination
I changed the way pagination works so it takes the id of the last
element, rather than an offset. I also added controller code to ask
for more elements than the client requested, so I can tell if there's
another page after the one requested. And, I added the nextPage field
to the view, so the client could grab it when needed.