adds validators and moves autoincrement to them
Validators are used to serialize creates (and, later, updates) with
the intention of being able to support upcerts later. Now creates are
synchronously run through validators to add an autoincrement id and
then asynchronously committed to the event log. You need to serialize
creates, but at the table level, so multiple creates to different
tables can happen concurrently. You still have to serialize them
globally when you commit them, but you already have the id so you
don't have to wait around for the commit to finish.