adds event for revalidating a table
In order fix the bug where the auto-increment wasn't being initialized
properly, I added an event for revalidating a table. This adds a
revalidate method to the validator protocol and an implementation for
AutoIncrement. It has the table call the validator with all the
entries and overwrites the entries with what it returns.

The AutoIncrement revalidate call goes through all the entries and
overwrites the id with the index.

I don't do anything for Upsert, since I don't need it,
and it's way more complicated, since I might have to rename tables.

It's not worth putting an interface on this now, so you need to commit
the event manually at the console.